Move files in the appointment book original project to the "apptbook" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
