+++ projects-parent/originals-parent/apptbook/src/it/java/edu/pdx/cs410J/apptbook/Project1IT.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.apptbook;
 
 import edu.pdx.cs410J.InvokeMainTestCase;
 import org.junit.Test;
