+++ projects-parent/originals-parent/apptbook/src/test/java/edu/pdx/cs410J/apptbook/AppointmentTest.java
@@ -1,10 +1,8 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.apptbook;
 
 import org.junit.Test;
 
-import static org.hamcrest.CoreMatchers.containsString;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.nullValue;
+import static org.hamcrest.CoreMatchers.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 
 /**
